authorization_service-master/requirements.txt
aiohttp==3.6.3
aioredis==1.3.1
aioprometheus[aiohttp,binary]
PyJWT==1.7.1